aws/etl/requirements.txt
requests
boto3
python-dotenv
sqlalchemy
psycopg2-binary